Add deploy-targeted HTTP E2E and fix form int coercion bugs
Introduce scripts/deploy-e2e.php for smoke/customer/full runs against
DEPLOY_BASE_URL with CSRF+Inertia handling and safe production guards.
Cast cart quantity and payment_method_id to int so form posts do not
TypeError in production-like HTTP clients.

+++ app/Http/Controllers/Shop/CartController.php
@@ -50,7 +50,11 @@ final class CartController extends Controller
             : null;
 
         try {
-            resolve(AddToCart::class)->handle($product, $variant, $data['quantity'] ?? 1);
+            resolve(AddToCart::class)->handle(
+                $product,
+                $variant,
+                (int) ($data['quantity'] ?? 1),
+            );
         } catch (InsufficientStockException) {
             Inertia::flash('toast', [
                 'type' => 'error',
